+++ apps/web/components/booking/actions/BookingActionsDropdown.tsx
@@ -32,6 +32,7 @@ import { ReportBookingDialog } from "@components/dialog/ReportBookingDialog";
 import { RerouteDialog } from "@components/dialog/RerouteDialog";
 import { RescheduleDialog } from "@components/dialog/RescheduleDialog";
 import { WrongAssignmentDialog } from "@components/dialog/WrongAssignmentDialog";
+import { RoutingTraceSheet } from "../RoutingTraceSheet";
 
 import { useBookingConfirmation } from "../hooks/useBookingConfirmation";
 import type { BookingItemProps } from "../types";
@@ -129,6 +130,10 @@ export function BookingActionsDropdown({
   const setRerouteDialogIsOpen = useBookingActionsStoreContext((state) => state.setRerouteDialogIsOpen);
   const isCancelDialogOpen = useBookingActionsStoreContext((state) => state.isCancelDialogOpen);
   const setIsCancelDialogOpen = useBookingActionsStoreContext((state) => state.setIsCancelDialogOpen);
+  const isOpenRoutingTraceSheet = useBookingActionsStoreContext((state) => state.isOpenRoutingTraceSheet);
+  const setIsOpenRoutingTraceSheet = useBookingActionsStoreContext(
+    (state) => state.setIsOpenRoutingTraceSheet
+  );
 
   const cardCharged = booking?.payment[0]?.success;
 
@@ -448,16 +453,23 @@ export function BookingActionsDropdown({
         status={getBookingStatus()}
       />
       {isBookingFromRoutingForm && (
-        <WrongAssignmentDialog
-          isOpenDialog={isOpenWrongAssignmentDialog}
-          setIsOpenDialog={setIsOpenWrongAssignmentDialog}
-          bookingUid={booking.uid}
-          routingReason={booking.assignmentReason[0]?.reasonString ?? null}
-          guestEmail={booking.attendees[0]?.email ?? ""}
-          hostEmail={booking.user?.email ?? ""}
-          hostName={booking.user?.name ?? null}
-          teamId={booking.eventType?.team?.id ?? null}
-        />
+        <>
+          <WrongAssignmentDialog
+            isOpenDialog={isOpenWrongAssignmentDialog}
+            setIsOpenDialog={setIsOpenWrongAssignmentDialog}
+            bookingUid={booking.uid}
+            routingReason={booking.assignmentReason[0]?.reasonString ?? null}
+            guestEmail={booking.attendees[0]?.email ?? ""}
+            hostEmail={booking.user?.email ?? ""}
+            hostName={booking.user?.name ?? null}
+            teamId={booking.eventType?.team?.id ?? null}
+          />
+          <RoutingTraceSheet
+            isOpen={isOpenRoutingTraceSheet}
+            setIsOpen={setIsOpenRoutingTraceSheet}
+            bookingUid={booking.uid}
+          />
+        </>
       )}
       {booking.paid && booking.payment[0] && (
         <ChargeCardDialog
@@ -703,19 +715,33 @@ export function BookingActionsDropdown({
                 </DropdownItem>
               </DropdownMenuItem>
               {isBookingFromRoutingForm && (
-                <DropdownMenuItem className="rounded-lg" key="report_wrong_assignment">
-                  <DropdownItem
-                    type="button"
-                    color="destructive"
-                    StartIcon="user-x"
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      setIsOpenWrongAssignmentDialog(true);
-                    }}
-                    data-testid="report_wrong_assignment">
-                    {t("report_wrong_assignment")}
-                  </DropdownItem>
-                </DropdownMenuItem>
+                <>
+                  <DropdownMenuItem className="rounded-lg" key="report_wrong_assignment">
+                    <DropdownItem
+                      type="button"
+                      color="destructive"
+                      StartIcon="user-x"
+                      onClick={(e) => {
+                        e.stopPropagation();
+                        setIsOpenWrongAssignmentDialog(true);
+                      }}
+                      data-testid="report_wrong_assignment">
+                      {t("report_wrong_assignment")}
+                    </DropdownItem>
+                  </DropdownMenuItem>
+                  <DropdownMenuItem className="rounded-lg" key="view_routing_trace">
+                    <DropdownItem
+                      type="button"
+                      StartIcon="route"
+                      onClick={(e) => {
+                        e.stopPropagation();
+                        setIsOpenRoutingTraceSheet(true);
+                      }}
+                      data-testid="view_routing_trace">
+                      {t("routing_trace")}
+                    </DropdownItem>
+                  </DropdownMenuItem>
+                </>
               )}
             </>
             <DropdownMenuSeparator />
